Added people mines which the sweepers avoid, implemented moving average for easier graph reading

+++ src/logic.hpp
@@ -26,7 +26,7 @@ class Logic
 		vector<Mine> mines;
 		int ticks;
 		int ticks2;
-		int max_fitness;
+		float max_fitness;
 
 		Control_Sweeper* control_sweeper = NULL;
 		Sweeper* best_sweeper = NULL;
